# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Reality Compiler API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/orders/{id}/refund:
    post:
      operationId: refundOrder
      tags: [orders]
      summary: Refund (full or partial) a paid order
      description: |
        Admin-only. Authenticated via the `x-admin-token` header against
        the server-side `ADMIN_API_TOKEN`. Triggers a Stripe Refund with
        `reverse_transfer` + `refund_application_fee` so the designer's
        Connect transfer and the platform fee are clawed back
        proportionally. The order's `paymentStatus` and `refundedAmount`
        columns are updated by the `charge.refunded` webhook.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: header
          name: x-admin-token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderRefundInput"
      responses:
        "200":
          description: Refund accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderRefundResult"
        "400":
          description: Invalid refund request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid admin token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/purge-deleted:
    post:
      operationId: purgeDeletedAccounts
      tags: [me]
      summary: Hard-delete accounts soft-deleted >30 days ago
      description: |
        Authenticated via the `x-admin-token` header against the
        server-side `ADMIN_API_TOKEN` env var. Intended to be invoked
        by a scheduled job.
      parameters:
        - in: header
          name: x-admin-token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Purge summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurgeSummary"
        "401":
          description: Invalid admin token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    PurgeSummary:
      type: object
      required: [usersPurged, sessionsPurged, listingsPurged, objectsDeleted]
      properties:
        usersPurged: { type: integer }
        sessionsPurged: { type: integer }
        listingsPurged: { type: integer }
        objectsDeleted: { type: integer }
...
    OrderRefundInput:
      type: object
      properties:
        amount:
          description: |
            Dollar amount to refund. Omit (or null) for a full refund of
            the remaining balance.
          type: ["number", "null"]
        reason:
          type: string
          enum: [duplicate, fraudulent, requested_by_customer]
    OrderRefundResult:
      type: object
      properties:
        orderId:
          type: integer
        refundId:
          type: string
        amount:
          type: number
        status:
          type: ["string", "null"]
      required:
        - orderId
        - refundId
        - amount